toolkit/components/payments/test/mochitest/test_rich_select.html

<!DOCTYPE HTML>
<html>
<!--
Test the rich-select component
-->
<head>
  <meta charset="utf-8">
  <title>Test the rich-select component</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="payments_common.js"></script>
  <script src="custom-elements.min.js"></script>
  <script src="ObservedPropertiesMixin.js"></script>
  <script src="rich-select.js"></script>
  <script src="rich-option.js"></script>
  <script src="address-option.js"></script>
  <script src="basic-card-option.js"></script>
  <link rel="stylesheet" type="text/css" href="rich-select.css"/>
  <link rel="stylesheet" type="text/css" href="address-option.css"/>
  <link rel="stylesheet" type="text/css" href="basic-card-option.css"/>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display">
    <rich-select id="select1">
      <address-option id="option1"
                      email="emzembrano92@email.com"
                      recipient="Emily Zembrano"
                      addressLine="717 Hyde Street #6"
                      city="San Francisco"
                      region="CA"
                      phone="415 203 0845"
                      postalCode="94109"
                      country="USA"></address-option>
      <address-option id="option2"
                      email="jenz9382@email.com"
                      recipient="Jennifer Zembrano"
                      addressLine="42 Fairydust Lane"
                      city="Lala Land"
                      region="HI"
                      phone="415 439 2827"
                      postalCode="98765"
                      country="USA"></address-option>
      <address-option id="option3"
                      email="johnz9382@email.com"
                      recipient="John Zembrano"
                      addressLine="42 Fairydust Lane"
                      city="Lala Land"
                      missinginformation="true"
                      region="HI"
                      phone="415 439 2827"
                      postalCode="98765"
                      country="USA"></address-option>
    </rich-select>

    <rich-select id="select2">
      <basic-card-option owner="Jared Wein"
                         expiration="01/1970"
                         number="4024007197293599"
                         type="Visa"></basic-card-option>
      <basic-card-option owner="Whimsy Corn"
                         expiration="01/1970"
                         number="5220465104517667"
                         type="Mastercard"></basic-card-option>
      <basic-card-option owner="Fire Fox"
                         expiration="01/1970"
                         number="6011777095481054"
                         type="Discover"></basic-card-option>
    </rich-select>
  </p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
<script type="application/javascript">
/** Test the rich-select address-option component **/

/* import-globals-from payments_common.js */
/* import-globals-from ../../res/components/address-option.js */
/* import-globals-from ../../res/components/basic-card-option.js */

let select1 = document.getElementById("select1");
let option1 = document.getElementById("option1");
let option2 = document.getElementById("option2");
let option3 = document.getElementById("option3");

function get_selected_clone() {
  return select1.querySelector(".rich-select-selected-clone");
}

function is_visible(element, message) {
  ok(!isHidden(element), message);
}

function is_hidden(element, message) {
  ok(isHidden(element), message);
}

function dispatchKeyDown(key, keyCode) {
  select1.dispatchEvent(new KeyboardEvent("keydown", {key, keyCode}));
}

add_task(async function test_addressLine_combines_address_city_region_postalCode_country() {
  ok(option1, "option1 exists");
  let addressLine = option1.querySelector(".addressLine");
  /* eslint-disable max-len */
  is(addressLine.textContent,
     `${option1.addressLine} ${option1.city} ${option1.region} ${option1.postalCode} ${option1.country}`);
  /* eslint-enable max-len */
});

add_task(async function test_no_option_selected_first_displayed() {
  ok(select1, "select1 exists");

  await asyncElementRendered();

  is_hidden(option1, "option 1 should be hidden when popup is not open");
  is_hidden(option2, "option 2 should be hidden when popup is not open");
  is_hidden(option3, "option 3 should be hidden when popup is not open");
  ok(option1.selected, "option 1 should be selected");
  ok(option1.hasAttribute("selected"), "option 1 should have selected attribute");
  let selectedClone = get_selected_clone();
  is_visible(selectedClone, "The selected clone should be visible at all times");
  is(selectedClone.getAttribute("email"), option1.getAttribute("email"),
     "The selected clone email should be equivalent to the selected option 1");
  is(selectedClone.getAttribute("recipient"), option1.getAttribute("recipient"),
     "The selected clone recipient should be equivalent to the selected option 1");
});

add_task(async function test_clicking_on_select_shows_all_options() {
  ok(select1, "select1 exists");
  ok(!select1.open, "select is not open by default");
  ok(option1.selected, "option 1 should be selected by default");

  select1.click();

  ok(select1.open, "select is open after clicking on it");
  ok(option1.selected, "option 1 should be selected when open");
  is_visible(option1, "option 1 is visible when select is open");
  is_visible(option2, "option 2 is visible when select is open");
  is_visible(option3, "option 3 is visible when select is open");

  option2.click();

  ok(!select1.open, "select is not open after blur");
  ok(!option1.selected, "option 1 is not selected after click on option 2");
  ok(option2.selected, "option 2 is selected after clicking on it");
  is_hidden(option1, "option 1 is hidden when select is closed");
  is_hidden(option2, "option 2 is hidden when select is closed");
  is_hidden(option3, "option 3 is hidden when select is closed");

  await asyncElementRendered();

  let selectedClone = get_selected_clone();
  is_visible(selectedClone, "The selected clone should be visible at all times");
  is(selectedClone.getAttribute("email"), option2.getAttribute("email"),
     "The selected clone email should be equivalent to the selected option 2");
  is(selectedClone.getAttribute("recipient"), option2.getAttribute("recipient"),
     "The selected clone recipient should be equivalent to the selected option 2");
});

add_task(async function test_changing_option_selected_affects_other_options() {
  ok(option2.selected, "Option 2 should be selected from prior test");

  option1.selected = true;
  ok(!option2.selected, "Option 2 should no longer be selected after making option 1 selected");
  ok(option1.hasAttribute("selected"), "Option 1 should now have selected attribute");
});

add_task(async function test_up_down_keys_change_selected_item() {
  let openObserver = new MutationObserver(mutations => {
    for (let mutation of mutations) {
      ok(mutation.attributeName != "open", "the select should not open/close during this test");
    }
  });
  openObserver.observe(select1, {attributes: true});

  ok(select1, "select1 exists");
  ok(option1.selected, "option 1 should be selected by default");

  ok(!select1.open, "select should not be open before focusing");
  select1.focus();
  ok(!select1.open, "select should not be open after focusing");

  dispatchKeyDown("ArrowDown", 40);
  ok(!option1.selected, "option 1 should no longer be selected");
  ok(option2.selected, "option 2 should now be selected");

  dispatchKeyDown("ArrowDown", 40);
  ok(!option2.selected, "option 2 should no longer be selected");
  ok(option3.selected, "option 3 should now be selected");

  dispatchKeyDown("ArrowDown", 40);
  ok(option3.selected, "option 3 should remain selected");
  ok(!option1.selected, "option 1 should not be selected");

  dispatchKeyDown("ArrowUp", 38);
  ok(!option3.selected, "option 3 should no longer be selected");
  ok(option2.selected, "option 2 should now be selected");

  dispatchKeyDown("ArrowUp", 38);
  ok(!option2.selected, "option 2 should no longer be selected");
  ok(option1.selected, "option 1 should now be selected");

  dispatchKeyDown("ArrowUp", 38);
  ok(option1.selected, "option 1 should remain selected");
  ok(!option3.selected, "option 3 should not be selected");

  // Wait for any mutation observer notifications to fire before exiting.
  await Promise.resolve();

  openObserver.disconnect();
});

add_task(async function test_open_close_from_keyboard() {
  select1.focus();

  ok(!select1.open, "select should not be open by default");

  dispatchKeyDown(" ", 32);
  ok(select1.open, "select should now be open");
  ok(option1.selected, "option 1 should be selected by default");

  dispatchKeyDown("ArrowDown", 40);
  ok(!option1.selected, "option 1 should not be selected");
  ok(option2.selected, "option 2 should now be selected");
  ok(select1.open, "select should remain open");

  dispatchKeyDown("ArrowUp", 38);
  ok(option1.selected, "option 1 should now be selected");
  ok(!option2.selected, "option 2 should not be selected");
  ok(select1.open, "select should remain open");

  dispatchKeyDown("Enter", 13);
  ok(option1.selected, "option 1 should now be selected");
  ok(!select1.open, "select should be closed");

  dispatchKeyDown(" ", 32);
  ok(select1.open, "select should now be open");

  dispatchKeyDown("Escape", 27);
  ok(!select1.open, "select should be closed");
});

add_task(async function test_clicking_on_options_maintain_one_item_always_selected() {
  ok(!select1.open, "select should be closed by default");
  ok(option1.selected, "option 1 should be selected by default");
  select1.click();
  ok(select1.open, "select should now be open");

  option3.click();
  ok(!select1.open, "select should be closed");
  ok(!option1.selected, "option 1 should be unselected");
  ok(option3.selected, "option 3 should be selected");

  select1.click();
  ok(select1.open, "select should open");
  ok(!option1.selected, "option 1 should be unselected");
  ok(option3.selected, "option 3 should be selected");

  option1.click();
  ok(!select1.open, "select should be closed");
  ok(option1.selected, "option 1 should be selected");
  ok(!option3.selected, "option 3 should be unselected");
});

add_task(async function test_selected_clone_should_equal_selected_option() {
  ok(option1.selected, "option 1 should be selected");
  await asyncElementRendered();

  let clonedOptions = select1.querySelectorAll(".rich-select-selected-clone");
  is(clonedOptions.length, 1, "there should only be one cloned option");

  let clonedOption = clonedOptions[0];
  for (let attrName of AddressOption.observedAttributes) {
    is(clonedOption.attributes[attrName] && clonedOption.attributes[attrName].value,
       option1.attributes[attrName] && option1.attributes[attrName].value,
       "attributes should have matching value; name=" + attrName);
  }

  option2.selected = true;
  await asyncElementRendered();

  clonedOptions = select1.querySelectorAll(".rich-select-selected-clone");
  is(clonedOptions.length, 1, "there should only be one cloned option");

  clonedOption = clonedOptions[0];
  for (let attrName of AddressOption.observedAttributes) {
    is(clonedOption.attributes[attrName] && clonedOption.attributes[attrName].value,
       option2.attributes[attrName] && option2.attributes[attrName].value,
       "attributes should have matching value; name=" + attrName);
  }
});

add_task(async function test_basic_card_simple() {
  let select2 = document.getElementById("select2");
  ok(select2, "basic card select should exist");
  let selectPopupBox = select2.querySelector(".rich-select-popup-box");
  ok(selectPopupBox, "basic card popup box exists");

  is(selectPopupBox.childElementCount, 3, "There should be three children in the popup box");

  let clonedOption = select2.querySelector(".rich-select-selected-clone");
  let selectedOption = selectPopupBox.firstChild;
  for (let attrName of BasicCardOption.observedAttributes) {
    is(clonedOption.attributes[attrName] && clonedOption.attributes[attrName].value,
       selectedOption.attributes[attrName] && selectedOption.attributes[attrName].value,
       "attributes should have matching value; name=" + attrName);
  }
});

</script>

</body>
</html>
